Start refactoring code to support running in background and adding multiple torrents
