+++ crates/librqbit/Cargo.toml
@@ -27,6 +27,7 @@ serde = {version = "1", features=["derive"]}
 serde_json = "1"
 anyhow = "1"
 
+regex = "1"
 reqwest = "0.11"
 urlencoding = "1"
 byteorder = "1"
